requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
httpx==0.25.0
python-dotenv==1.0.0
pydantic==1.10.12
joblib==1.3.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
pytest==7.4.0
pytest-asyncio==0.21.1